requirements.txt
asgiref==3.8.1
attrs==25.1.0
channels==4.2.0
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
django-modeltranslation==0.19.12
django-unfold==0.46.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
